I have tried adding in requests for data at set intervals currently at 10 seconds, I would like to go over this more and make sure it is working, however it seems to be working ok at the moment

+++ src/token.rs
@@ -58,7 +58,7 @@ impl TokenManager {
             .unwrap()
             .as_secs_f64();
 
-        if now >= self.time_token_was_made + 1770 as f64 {
+        if now >= self.time_token_was_made + 1770_f64 {
             self.update_token().await.unwrap()
         } else {
             self.token.clone() // probably want to avoid clone
